+++ HISTORY.rst
@@ -1,3 +1,12 @@
+0.3.0 (2016-11-24)
+++++++++++++++++++
+
+- Added multi-lines graph-frames
+- Deprecated the usage of "_init" and "_update"
+- Added new decorator "callit"
+- Added the possibility to use ax-related kwargs in graph-frames
+- Allowed xnpts and xnptsmax = None to apply no limit on the amount data plotted
+
 0.1.4 (2016-10-14)
 ++++++++++++++++++